# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ConfigRequest": {
        "additionalProperties": false,
        "properties": {
          "atlas_space": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Atlas template name, must match source when atlas is supplied.",
            "title": "Atlas Space"
          },
          "confound_columns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Columns to regress; null selects complete motion6 + available WM/CSF.",
            "title": "Confound Columns"
          },
          "data_space": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Exact source template name. All image inputs require a known space.",
            "title": "Data Space"
          },
          "detrend": {
            "default": true,
            "description": "Remove linear trends through Nilearn signal.clean.",
            "title": "Detrend",
            "type": "boolean"
          },
          "discard": {
            "default": 0,
            "description": "Number of initial original frames to remove, integer >=0.",
            "title": "Discard",
            "type": "integer"
          },
          "fd_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Censor FD > cutoff in mm, also undefined first FD; null disables FD censoring.",
            "title": "Fd Threshold"
          },
          "high_pass": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "High-pass cutoff in Hz; null disables. Requires TR and cutoff below Nyquist.",
            "title": "High Pass"
          },
          "low_pass": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Low-pass cutoff in Hz; null disables. Must exceed high_pass when both are set.",
            "title": "Low Pass"
          },
          "method": {
            "default": "pearson",
            "description": "pearson, spearman, or partial (Ledoit-Wolf shrinkage precision).",
            "title": "Method",
            "type": "string"
          },
          "min_samples": {
            "default": 20,
            "description": "Minimum retained frames (>=3), not a statistical power criterion.",
            "title": "Min Samples",
            "type": "integer"
          },
          "preprocessed": {
            "default": false,
            "description": "Explicit declaration that image motion correction/registration are complete.",
            "title": "Preprocessed",
            "type": "boolean"
          },
          "standardize": {
            "default": true,
            "description": "Sample z-score of cleaned ROI time series (ddof=1).",
            "title": "Standardize",
            "type": "boolean"
          },
          "t_r": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Repetition time in seconds; null infers from scan JSON/header; conflicts are rejected.",
            "title": "T R"
          },
          "table_header": {
            "default": true,
            "description": "Text first non-comment row contains ROI names; ignored for NPY/NPZ/MAT.",
            "title": "Table Header",
            "type": "boolean"
          },
          "transpose": {
            "default": false,
            "description": "Transpose table input to time x ROI, discarding text column names.",
            "title": "Transpose",
            "type": "boolean"
          },
          "variable": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "NPZ/MAT 2D numeric variable name; null requires one unambiguous variable.",
            "title": "Variable"
          }
        },
        "title": "ConfigRequest",
        "type": "object"
      },
...
      "ExtractionRequest": {
        "additionalProperties": false,
        "description": "One source run, optional companion paths and scientific settings.",
        "properties": {
          "atlas": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Integer-label atlas in the source space.",
            "title": "Atlas"
          },
          "config": {
            "$ref": "#/components/schemas/ConfigRequest"
          },
          "confounds": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Headered confounds TSV/CSV with all original frames.",
            "title": "Confounds"
          },
          "guidance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Guidance"
              },
              {
                "type": "null"
              }
            ]
          },
          "mask": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Same-space volume mask; positive voxels retained.",
            "title": "Mask"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Display-only same-space brain mask/skull-stripped T1.",
            "title": "Reference"
          },
          "rois": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ROI mapping CSV/TSV; exact IDs and optional RAS+ coordinates.",
            "title": "Rois"
          },
          "source": {
            "description": "Absolute or server-relative path to one fMRI/ROI file.",
            "minLength": 1,
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source"
        ],
        "title": "ExtractionRequest",
        "type": "object"
      },
      "Guidance": {
        "additionalProperties": true,
        "description": "Optional GUI confirmation record; programmatic callers may omit it.",
        "properties": {
          "confounds_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "upstream or skip when confounds absent.",
            "title": "Confounds Decision"
          },
          "dataset": {
            "description": "Catalog dataset id, e.g. custom or adni.",
            "title": "Dataset",
            "type": "string"
          },
          "denoise_confirmed": {
            "default": false,
            "title": "Denoise Confirmed",
            "type": "boolean"
          },
          "input_kind": {
            "default": "auto",
            "description": "Original route, such as auto or raw-bids.",
            "title": "Input Kind",
            "type": "string"
          },
          "raw_qc_confirmed": {
            "default": false,
            "title": "Raw Qc Confirmed",
            "type": "boolean"
          },
          "reference_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "skip when reference absent.",
            "title": "Reference Decision"
          },
          "review_confirmed": {
            "default": false,
            "title": "Review Confirmed",
            "type": "boolean"
          },
          "source_confirmed": {
            "default": false,
            "title": "Source Confirmed",
            "type": "boolean"
          },
          "spatial_confirmed": {
            "default": false,
            "title": "Spatial Confirmed",
            "type": "boolean"
          },
          "variant": {
            "description": "Exact variant id returned by /api/presets.",
            "title": "Variant",
            "type": "string"
          }
        },
        "required": [
          "dataset",
          "variant"
        ],
        "title": "Guidance",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobState": {
        "additionalProperties": true,
        "description": "Persistent job state, returned immediately on submission and during polling.",
        "properties": {
          "atlas": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atlas"
          },
          "error_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Type"
          },
          "example_kind": {
            "anyOf": [
              {
                "enum": [
                  "rest01",
                  "synthetic"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Example identity for demo jobs; old jobs may omit this field.",
            "title": "Example Kind"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "extract",
              "demo",
              "atlas",
              "dicom",
              "preprocess",
              "python-preprocess",
              "python-dicom"
            ],
            "title": "Kind",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "qc": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qc"
          },
          "result_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Dir"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "complete",
              "failed",
              "interrupted"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "message",
          "kind"
        ],
        "title": "JobState",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local single-run fMRI connectivity API. Paths refer to this server's filesystem. Use /api/preflight before /api/jobs, then poll the returned id. Extraction shares the Python core. Bind to loopback and use one server per workspace. Responses and exports include local paths.",
    "title": "BrainFC",
    "version": "0.5.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/jobs": {
      "get": {
        "operationId": "jobs_api_jobs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobState"
                  },
                  "title": "Response Jobs Api Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List up to 100 jobs, most recently modified first",
        "tags": [
          "Jobs"
        ]
      },
      "post": {
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue one extraction; does not wait for completion",
        "tags": [
          "Jobs"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "operationId": "job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Poll persisted job state",
        "tags": [
          "Jobs"
        ]
      }
    },
...
  }
}